+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `--color=[(auto)|on|off]` command line flag
+
 ## [3.4.0] - 2024-08-28
 
 ### Added